requirements.txt
# Core dependencies
streamlit>=1.24.0
python-dotenv>=0.19.0
pydantic>=2.0.0

# LLM providers
openai>=1.0.0
anthropic>=0.3.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.12.0
mypy>=1.0.0

# Configuration managements
pydantic-settings>=2.0.0

